create-profmark now working, and has generated its first benchmark. concatdb no longer needed (sucked into create-profmark); deleted. [J3/16-19]
hmmsearch now shows a header, and some summary statistics at the end
of a run.

Updated to deal w/ easel's new esl_randomseq module -- sequence
sampling routines split away from esl_random.




git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/trunk@2400 db3e70e4-12e1-0310-b23f-81a6e3d09ba7

+++ src/hmmer.h
@@ -452,7 +452,7 @@ typedef struct p7_dom_s {
   float          seqcorrection;	/* null2 correction to add to per-seq null score for this domain  */
   float          domcorrection;	/* null2 correction to add null score when calculating a per-domain score */
   float          bitscore;	/* overall score in bits, null corrected, if this were the only domain in seq */
-  float          pvalue;	/* P-value of the bitscore */
+  double         pvalue;	/* P-value of the bitscore */
   P7_ALIDISPLAY *ad; 
 } P7_DOMAIN;
